requirements.txt
anthropic
python-telegram-bot
requests
schedule
